Refactor server utilities and types
