+++ app/lib/colorApi.server.ts
@@ -1,21 +1,8 @@
+import type { ColorDescriptor, GetColorByHslOptions } from "./types.server";
+import { normalizeHue } from "./utils.server";
+
 // Base endpoint documented by The Color API for looking up an HSL tuple.
 const COLOR_API_ENDPOINT = "https://www.thecolorapi.com/id";
-
-export interface ColorDescriptor {
-  readonly name: string;
-  readonly rgb: {
-    readonly value: string;
-    readonly r: number;
-    readonly g: number;
-    readonly b: number;
-  };
-  readonly hsl: {
-    readonly value: string;
-    readonly h: number;
-    readonly s: number;
-    readonly l: number;
-  };
-}
 
 interface ColorApiResponse {
   readonly name: { readonly value: string };
@@ -33,12 +20,6 @@ interface ColorApiResponse {
   };
 }
 
-export interface GetColorByHslOptions {
-  readonly hue: number;
-  readonly saturation: number;
-  readonly lightness: number;
-}
-
 // Guard against out-of-band saturation/lightness values before building the URL.
 const clampPercentage = (value: number): number => {
   if (Number.isNaN(value)) {
@@ -54,20 +35,6 @@ const clampPercentage = (value: number): number => {
   }
 
   return value;
-};
-
-// Convert arbitrary hue input into the `[0, 360)` range expected by the API.
-const normalizeHue = (hue: number): number => {
-  if (!Number.isFinite(hue)) {
-    return 0;
-  }
-
-  const wrapped = hue % 360;
-  if (wrapped < 0) {
-    return wrapped + 360;
-  }
-
-  return wrapped;
 };
 
 // Query The Color API for a single color description at the provided HSL triplet.
@@ -115,4 +82,3 @@ export async function getColorByHsl({
   } satisfies ColorDescriptor;
 }
 
-export type { ColorDescriptor as ColorSample };
